fix int32 issue with sector offset

+++ web/App_Code/BLL/DynamicClientPartition/ClientPartition.cs
@@ -127,7 +127,7 @@ namespace BLL.DynamicClientPartition
                     partCounter++;
 
                     //Determine what sector the first partition should start at
-                    if (Convert.ToInt32(schemaPartition.Start) < FirstPartitionStartSector)
+                    if (Convert.ToInt64(schemaPartition.Start) < FirstPartitionStartSector)
                         FirstPartitionStartSector = Convert.ToInt32(schemaPartition.Start);
 
                     if (!schemaPartition.Active)
